+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 ## 12.x.x (unreleased)
 
 ### Improvements
-- Client: Preferences Registry
 
 ### Bugfixes
 - 
@@ -10,10 +9,17 @@
 You should also include the user name that made the change.
 -->
 
-## 12.x.x (unreleased)
+## 12.118.1 (2022/08/08)
+
+### Bugfixes
+- Client: can not show some setting pages @syuilo
+
+## 12.118.0 (2022/08/07)
 
 ### Improvements
+- Client: 設定のバックアップ/リストア機能
 - Client: Add vi-VN language support
+- Client: Add unix time widget @syuilo
 
 ### Bugfixes
 - Server: リモートユーザーを正しくブロックできるように修正する @xianonn